edanalytica/requirements.txt
django
mysqlclient==1.4.2.post1
Pillow
Jinja2
djangorestframework